requirements.txt
requests
telethon
flask
python-dotenv